Added function to extract out public IP of the machine.
This proved to be a bit tricky. gethostbyname and getaddrinfo are not useful here since they will return
the Loopback IP address itself. We use 'getifaddrs' which returns a linked list of all interfaces and when
we encouter first interface with public IP and which is UP, we use that interface.

Addition of peer_info_t struct.

Signed-off-by: Salil <slashforu@gmail.com>